Derive Macro rithmic::CmpByKey

#[derive(CmpByKey)]
{
    // Attributes available to this derive:
    #[key]
}

Derive PartialEq, Eq, PartialOrd and Ord for a struct, but only based on a subset of its fields, marked by #[key]

This is particularly useful on structs where some of the other fields do not implement these traits.

Examples

use rithmic::CmpByKey;

#[derive(CmpByKey)]
struct City {
    loc: (f64, f64),
    #[key] pop: u64
}

let a = [
    City { loc: (35.689722, 139.692222), pop: 13_988_129 },
    City { loc: (53.763056, -122.745278), pop: 74_003 },
    City { loc: (51.507222, -0.12750000), pop: 9_002_488 },
];

assert!(a[0] > a[2]);
assert!(a[1] < a[2]);