Function rithmic::mo_algorithm

pub fn mo_algorithm(
    n: usize, 
    queries: impl IntoIterator<Item = impl Rangelike<usize>>
) -> impl Iterator<Item = MoStep>

Mo’s Algorithm

Given a list of Q range queries about a sequence of length N, produce an iterator of instructions MoStep::Add, Remove, Answer that answers all queries in O((N+Q) √N) steps

Examples

let seq = [10000, 2000, 300, 40, 5];
let queries = [2..5, 0..3, 1..2];
let mut answers = [0; 3];
let mut current = 0;

for step in mo_algorithm(seq.len(), queries) {
match step {
    MoStep::Add(i)    => current += seq[i],
    MoStep::Remove(i) => current -= seq[i],
    MoStep::Answer(q) => answers[q] = current
}}

assert_eq!(answers, [345, 12300, 2000]);