Struct rithmic::gameboard::Gameboard

pub struct Gameboard {
    pub h: usize,
    pub w: usize,
    pub blocked: Option<BVec>,
}

The Gameboard struct represents a rectangular grid of squares, some of which may be “blocked”, and provides convenience methods for determining what squares can be reached by various movement patterns

Fields
h: usizew: usizeblocked: Option<BVec>
Implementations

impl Gameboard

pub fn new(h: usize, w: usize) -> Self

Create a new Gameboard with h rows and w columns. No squares are blocked initially

pub fn from_vec_vec<T: PartialEq>(blocked: Vec<Vec<T>>, block: T) -> Self

Create a new Gameboard from a Vec<Vec<T>>. Any blocked[i][j] == block means [i, j] is blocked; any other values are ignored

pub fn inbounds(&self, [i, j]: [usize; 2]) -> bool

Check if [i, j] is within the board dimensions. Blocks are not checked

pub fn block(&mut self, [i, j]: [usize; 2])

Set a block at [i, j]. No effect if already blocked

pub fn unblock(&mut self, [i, j]: [usize; 2])

Remove any block at [i, j]. No effect is there is no block there

pub fn is_blocked(&self, [i, j]: [usize; 2]) -> bool

Check if there is currently a block at [i, j]

pub fn dir_jump(
&self,
[i, j]: [usize; 2],
dir: u8,
dist: usize
) -> Option<[usize; 2]>

Move dist squares in a straight line. Blocks along the way are ignored, but a block at the destination, or the destination being out-of-bounds, returns None. dir must be in 0..4 and indicates direction:

Positive i

Positive j

Negative i

Negative j

pub fn diag_jump(
&self,
[i, j]: [usize; 2],
dir: u8,
dist: usize
) -> Option<[usize; 2]>

Move dist squares in a 45° diagonal line. Blocks along the way are ignored, but a block at the destination, or the destination being out-of-bounds, returns None. dir must be in 0..4 and indicates direction:

Positive i and j

Negative i, positive j

Negative i and j

Positive i, negative j

pub fn dpad4(&self, [i, j]: [usize; 2]) -> ArrayVec<[[usize; 2]; 4]>

Of the four squares orthogonally adjacent to [i, j], return those that are within the board bounds and not blocked

pub fn dpad8(&self, [i, j]: [usize; 2]) -> ArrayVec<[[usize; 2]; 8]>

Of the eight squares surrounding [i, j] (orthogonally and diagonally), return those that are within the board bounds and not blocked

pub fn knight_jumps(&self, [i, j]: [usize; 2]) -> ArrayVec<[[usize; 2]; 8]>

Of the squares reachable by “knight’s move” from [i, j], return those that are within the board bounds and not blocked

pub fn rook_moves(&self, [i, j]: [usize; 2]) -> Vec<[usize; 2]>

Of the squares reachable by “rook’s move” from [i, j], return those that are within the board bounds and not blocked

pub fn bishop_moves(&self, [i, j]: [usize; 2]) -> Vec<[usize; 2]>

Of the squares reachable by “bishop’s move” from [i, j], return those that are within the board bounds and not blocked

pub fn queen_moves(&self, [i, j]: [usize; 2]) -> Vec<[usize; 2]>

Of the squares reachable by “queen’s move” from [i, j], return those that are within the board bounds and not blocked

pub fn manhat_jumps(&self, [i, j]: [usize; 2], dist: usize) -> Vec<[usize; 2]>

Of the squares that are dist Manhattan distance away from [i, j], return those that are within the board bounds and not blocked