Struct rithmic::graph::prelude::Graph

pub struct Graph<E, const FLAGS: usize>where
    E: Copy,{
    pub adj: Vec<Vec<(usize, E)>>,
}

Fields
adj: Vec<Vec<(usize, E)>>
Implementations

impl<E, const FLAGS: usize> Graph<E, FLAGS>where E: Copy,

pub fn new(size: usize) -> Self

pub fn add_edge(&mut self, u: usize, v: usize, edge: E)

pub fn from_iter( size: usize, iter: impl IntoIterator<Item = (usize, usize, E)> ) -> Self

pub fn size(&self) -> usize

pub fn rev(&self) -> Self

impl<const FLAGS: usize> Graph<(), FLAGS>

pub fn from_iter_unweighted( size: usize, iter: impl IntoIterator<Item = (usize, usize)> ) -> Self

pub fn map<const DEST_FLAGS: usize>( &self, dest_size: usize, f: impl FnMut(usize) -> usize ) -> Graph<(), DEST_FLAGS>

impl<E, const FLAGS: usize> Graph<E, FLAGS>where E: Copy,

pub fn dfs(&self, s: impl IntoIterator<Item = usize>) -> Vec<DfsStep<E>>

impl<E, const FLAGS: usize> Graph<E, FLAGS>where E: Copy + Default + Add<Output = E> + PartialOrd,

pub fn floyd_warshall(&self) -> NdVec<2, Option<(E, usize)>>

impl<const FLAGS: usize> Graph<(), FLAGS>

pub fn bfs( &self, s: impl IntoIterator<Item = usize>, t: impl IntoIterator<Item = usize> ) -> DistPred<usize>

impl<const FLAGS: usize> Graph<bool, FLAGS>

pub fn dijkstra_01( &self, s: impl IntoIterator<Item = usize>, t: impl IntoIterator<Item = usize> ) -> DistPred<usize>

impl<E, const FLAGS: usize> Graph<E, FLAGS>where E: Copy + Default + Add<Output = E> + Ord,

pub fn dijkstra( &self, s: impl IntoIterator<Item = usize>, t: impl IntoIterator<Item = usize> ) -> DistPred<E>

impl<E> Graph<E, TREE>where E: Copy,

pub fn lca(&self, s: usize) -> Lca

impl<E, const FLAGS: usize> Graph<E, FLAGS>where E: Copy,

pub fn tarjan_scc(&self) -> SCCs

impl<E> Graph<E, TREE>where E: Copy,

pub fn dfs_up_tree(&self, s: usize) -> impl Iterator<Item = (usize, usize, E)>

(Tree) Filter a DFS of the tree to only the n - 1 “upward” edges u → p where p is the parent of u in the DFS

Examples

let g = Tree::line(4);

assert_eq!(g.dfs_up_tree(0).collect::<Vec<_>>(), vec![
    (3, 2, ()),
    (2, 1, ()),
    (1, 0, ())
]);

pub fn euler_tour(&self, s: usize) -> impl Iterator<Item = usize>

(Tree) List node visits in a DFS starting at s

Examples

let g = Tree::line(4);

assert_eq!(g.euler_tour(0).collect::<Vec<_>>(), vec![0, 1, 2, 3, 2, 1, 0]);

pub fn euler_tour_edges( &self, s: usize ) -> impl Iterator<Item = (usize, usize, E)>

(Tree) List edges traversed in a DFS starting at s

Examples

let mut g = Tree::<&str>::new(3);
g.add_edge(0, 1, "Main Street");
g.add_edge(1, 2, "Privet Drive");

assert_eq!(g.euler_tour_edges(0).collect::<Vec<_>>(), vec![
    (0, 1, "Main Street"),
    (1, 2, "Privet Drive"),
    (2, 1, "Privet Drive"),
    (1, 0, "Main Street")
]);

pub fn subtree_sizes_rel_to(&self, root: usize) -> Vec<usize>

(Tree) Relative to root, determine the number of nodes descendant of node u (including u) for all u, in O(n)

Examples

A 3-pointed (n = 4) star graph, where node 0 is the center

let g = Tree::star(4);

assert_eq!(g.subtree_sizes_rel_to(0), vec![4, 1, 1, 1]);
assert_eq!(g.subtree_sizes_rel_to(1), vec![3, 4, 1, 1]);

pub fn subtree_sizes(&self) -> BTreeMap<(usize, usize), usize>

(Tree) Determine the size of all 2(n - 1) possible subtrees that can be created by deleting a single edge. See rooted_subtree_fn for illustration of the naming scheme

impl Graph<E, TREE>

pub fn star(n: usize) -> Self

pub fn line(n: usize) -> Self

impl<E> Graph<E, TREE>where E: Copy,

pub fn rooted_subtree_fn<R>( &self, f: impl FnMut(RsfStep<'_, E, R>) -> R ) -> BTreeMap<(usize, usize), R>where R: Clone,

O(n)-step calculation of function f() for:

All of the 2(n - 1) rooted subtrees induced by choosing an edge (p, u), rooting at u, and removing all vertices on the p side

graph TD 0(("0 (p)")) --- 1((1)) 0 -.- 2(("2 (u)")) 1 --- 3((3)) 2 === 4((4)) 2 === 5((5)) style 2 fill:#267,stroke:#bbb,stroke-width:4px style 4 fill:#267,stroke:#bbb,stroke-width:4px style 5 fill:#267,stroke:#bbb,stroke-width:4px

graph TD 0(("0 (u)")) === 1((1)) 0 -.- 2(("2 (p)")) 1 === 3((3)) 2 --- 4((4)) 2 --- 5((5)) style 0 fill:#267,stroke:#bbb,stroke-width:4px style 1 fill:#267,stroke:#bbb,stroke-width:4px style 3 fill:#267,stroke:#bbb,stroke-width:4px

All of the n trees rooted at some node u of the tree

graph TD 0((0)) === 1((1)) 0 === 2(("2 (u)")) 1 === 3((3)) 2 === 4((4)) 2 === 5((5)) style 0 fill:#267,stroke:#bbb,stroke-width:4px style 1 fill:#267,stroke:#bbb,stroke-width:4px style 2 fill:#267,stroke:#bbb,stroke-width:4px style 3 fill:#267,stroke:#bbb,stroke-width:4px style 4 fill:#267,stroke:#bbb,stroke-width:4px style 5 fill:#267,stroke:#bbb,stroke-width:4px

Requirements

Using very informal notation:

f() must be calculable for a leaf node

f() must have some way to efficiently “combine two sibling subtrees”. Let us call this operation ⨂

f() must have some way to efficiently “combine a subtree, with no siblings, with its parent”. Let us call this operation ⨁

Then for a parent p of siblings u, v, w, the idea is

f(p) = p ⨁ [ f(u) ⨂ f(v) ⨂ f(w) ]

API

f() will be passed RsfSteps:

f(RsfStep::Leaf { p, edge, u }) must return f(u) for a leaf node u

f(RsfStep::Sibling { p, r0, r1 }) must return r0 ⨂ r1

f(RsfStep::Parent { pp, edge, p, r }) must return p ⨁ r

When all steps are complete, the returned BTreeMap contains 3n - 2 entries:

[&(p, u)] stores f() for an induced subtree rooted at u and pruned at p as described above

[&(NONE, u)] stores f() for the full tree rooted at u as described above

Complexity

O(n log n), due to BTreeMap, assuming f() is O(1)

Examples

Calculate all subtree sizes in a 6-pointed (n = 7) star graph

use rithmic::graph::prelude::*;

let g = Tree::star(7);
let rsf = g.rooted_subtree_fn( |step| {
   match step {
       RsfStep::Leaf { .. } => 1,
       RsfStep::Sibling { r0, r1, .. } => *r0 + *r1,
       RsfStep::Parent { r, .. } => *r + 1,
   }
});

assert_eq!(rsf.len(), 2*6 + 7);
for u in 1..7 {
   assert_eq!(rsf[&(0, u)], 1);
   assert_eq!(rsf[&(u, 0)], 6);
}
for u in 0..7 {
   assert_eq!(rsf[&(NONE, u)], 7);
}

pub fn rooted_subtree_fn_half<R>( &self, f: impl FnMut(RsfStep<'_, E, R>) -> R ) -> BTreeMap<(usize, usize), R>where R: Clone,

Run rooted_subtree_fn but stop after the first upward DFS (normally there are two; one up then down). This may reduce runtime by more than half. The returned BTreeMap contains:

[&(p, u)] for the n - 1 induced subtrees where u is farther from 0 than p, but not the reverse

[&(NONE, 0)] for the full tree rooted at 0, but no other roots