Macro rithmic::struct_input

struct_input!() { /* proc-macro */ }

Given a ::proconio input specification and a Name, define:

A struct definition called Name

A macro called name_input! that call’s proconio’s input! with the given specification

A macro called name_destruct! that expands to a destructure of the Name struct and can be used to either construct from free variables, or destructure into free variables (see example below)

Note: #![feature(rustc_attrs)] and #![feature(decl_macro)] are required

Examples

#![feature(decl_macro, rustc_attrs)]

use rithmic::struct_input;

struct_input! {
    Student {
        name: String,
        exams_taken: usize,
        scores: [u8; exams_taken],
    }
}

// We now have access to:
// Student
// student_input!()
// student_destruct!() which expands to:
//      Student { name, exams_taken, scores }

// In a real program, this would read `name`, `exams_taken` and `scores` from stdin:
// student_input!();

assert_eq!(name, "Melissa");
assert_eq!(exams_taken, 3);
assert_eq!(scores, vec![85, 99, 71]);

// use _destruct! to construct from free variables
let x: Student = student_destruct!();

assert_eq!(x.name, "Melissa");
assert_eq!(x.exams_taken, 3);
assert_eq!(x.scores, vec![85, 99, 71]);

// use _destruct! to destruct to free variables
let student_destruct!() = x;

assert_eq!(name, "Melissa");
assert_eq!(exams_taken, 3);
assert_eq!(scores, vec![85, 99, 71]);