Module rithmic::prelude

Re-exports

pub use crate::AhoCorasick;

pub use crate::ConvexHullMax;

pub use crate::ConvexHullMin;

pub use crate::List;

pub use crate::SortedList;

pub use crate::binary_search;

pub use crate::binary_search;

pub use crate::par_binary_search;

pub use crate::BVec;

pub use crate::DebugOr;

pub use crate::DisjointIntervals;

pub use crate::Gameboard;

pub use crate::Vector2DMore;

pub use crate::imax;

pub use crate::imin;

pub use crate::IndexCompress;

pub use crate::IndexCompressed;

pub use crate::Insort;

pub use crate::IntBitOps;

pub use crate::CountInversions;

pub use crate::LongestTransitiveSubseq;

pub use crate::Ipc;

pub use crate::MeldHeap;

pub use crate::MeldMinHeap;

pub use crate::mo_algorithm;

pub use crate::mo_algorithm;

pub use crate::MoStep;

pub use crate::FactorialTable;

pub use crate::monoid_ops;

pub use crate::NdFenwick;

pub use crate::SegTree;

pub use crate::NdVec;

pub use crate::OdometerBE;

pub use crate::OdometerLE;

pub use crate::OptionMerge;

pub use crate::OrdPair;

pub use crate::polynomial_add;

pub use crate::polynomial_mul;

pub use crate::polynomial_div;

pub use crate::polynomial_div_exact;

pub use crate::lagrange_interpolation;

pub use crate::PrefixSums;

pub use crate::PrintMethods;

pub use crate::Rangelike;

pub use crate::Divisors;

pub use crate::GetSub;

pub use crate::SparseTable;

pub use crate::ternary_search;

pub use crate::ternary_search;

pub use crate::triangular_n;

pub use crate::triangular_n;

pub use crate::triangular_slice;

pub use crate::triangular_steps;

pub use crate::UnionFind;

pub use crate::UnwrapAny;

pub use crate::Waterline;

pub use crate::XorBasis128;

pub use crate::XorBasis64;

pub use crate::XorBasis32;

pub use crate::XorBasis16;

pub use crate::XorBasis8;

Macros

struct_input

Given a ::proconio input specification and a Name, define:

Attribute Macros

autofill

Create a wrapper macro with the same name as a given function, which automatically fills some of the later arguments with variables of the same name in the calling scope. The result is similar to a closure’s external variable capture but without the disadvantages of closures (e.g. lack of recursion, borrow conflicts)

memoize

Memoize a function based on some or all of its arguments

opaque

See transparent

semitransparent

See transparent

transparent

Change the hygiene of a declarative macro

Derive Macros

CmpByKey

Derive [PartialEq], [Eq], [PartialOrd] and [Ord] for a struct, but only based on a subset of its fields, marked by #[key]

CmpFromOrd