Struct rithmic::Ipc

pub struct Ipc {
    pub tx: ChildStdin,
    pub rx: BufReader<ChildStdout>,
    /* private fields */
}

Inter-process communication via pipes with a spawned child command

The child’s stdin & stdout can be accessed via the tx and rx fields. See also the input() and input_line() convenience methods. The child process is killed (if still alive) when this object is dropped

Examples

use std::io::Write;
use std::process::Command;
use proconio::input;

let mut ipc = Ipc::new(
    Command::new("python3")
        .arg("-c")
        .arg("print(int(input()) + 11)")
);

writeln!(&mut ipc.tx, "23");
input! { from ipc.input_line(), x: i32 };

assert_eq!(x, 34);

Fields
tx: ChildStdinrx: BufReader<ChildStdout>
Implementations

impl Ipc

pub fn new(cmd: &mut Command) -> Self

Spawn a child process and set up piped I/O to it. See Ipc for an example

pub fn input_line(&mut self) -> impl Source<&[u8]>

Wait for the next line from the child process, then create a ::proconio::source::Source from it. See Ipc for an example

pub fn input(&mut self) -> impl Source<&[u8]>

Read from child process until EOF (blocking), then create a ::proconio::source::Source from the read data