Trait rithmic::Rangelike

pub trait Rangelike<T> {
    fn clamp(&self, to: Range<T>) -> Option<(T, T)>;
    fn canonical(&self) -> (Option<T>, Option<T>);
}

This trait is intended to allow function APIs to interchangeably accept integers or any of their Range variants

2-tuples and 2-arrays are also accepted

Examples

let mut st = SegTree::<i32>::from([2, 3, 5, 7, 11]);

assert_eq!(st.query(2), 5);
assert_eq!(st.query(2..=3), 5+7);
assert_eq!(st.query(2..), 5+7+11);
assert_eq!(st.query(..), 2+3+5+7+11);

Required Methods

fn clamp(&self, to: Range<T>) -> Option<(T, T)>

Convert to (start_included, end_excluded) and check the range is a subset of to, otherwise return None. Unbounded side(s) are substituted with to.start or to.end respectively.

fn canonical(&self) -> (Option<T>, Option<T>)

Convert to (start_included, end_excluded) where each side is None if unbounded