Module rithmic::prelude

Re-exports

pub use crate::Gameboard;

pub use crate::MeldHeap;

pub use crate::MeldMinHeap;

pub use crate::monoid_ops;

pub use crate::polynomial_add;

pub use crate::polynomial_mul;

pub use crate::polynomial_div;

pub use crate::polynomial_div_exact;

pub use crate::lagrange_interpolation;

pub use crate::triangular_n;

pub use crate::triangular_n;

pub use crate::triangular_slice;

pub use crate::triangular_steps;

pub use crate::XorBasis128;

pub use crate::XorBasis64;

pub use crate::XorBasis32;

pub use crate::XorBasis16;

pub use crate::XorBasis8;

Macros

imax

Incremental maximization

imin

Incremental minimization

Structs

BVec

A compact boolean (bitset) vector with efficient << >> shifting and | ^ & logical operators

FactorialTable

IndexCompressed

This struct is created by IndexCompress::index_compress(). See its documentation for more

Ipc

Inter-process communication via pipes with a spawned child command

List

NdFenwick

NdVec

SegTree

SortedList

SparseTable

UnionFind

A union-find data structure

Enums

MoStep

This enum is created by mo_algorithm(). See its documentation for more

Traits

CountInversions

Divisors

Add a divisors method to ::primal::Sieve

GetSub

get_sub(index, subtract) indexes position [index - subtract] of a slice, but handles underflow (returning None)

IndexCompress

Index compress a sequence of length n by mapping every value to its ordered rank

Insort

IntBitOps

This trait provides methods for working with unsigned integers at the bit level

OptionMerge

OrdPair

Order a pair (2-tuple or 2-array)

PairMut

Mutably borrow two indexes of a slice simultaneously.

PrefixSums

Calculate prefix or suffix sums

PrintMethods

Enable functional-style printing

Rangelike

This trait is intended to allow function APIs to interchangeably accept integers or any of their Range variants

UnwrapAny

Unwrap a Result<T, T>, ignoring the Ok/Err status

Vector2DMore

Functions

binary_search

Binary search domain for the first point at which predicate returns true or false matching search_for

mo_algorithm

Mo’s Algorithm

ternary_search

Ternary search for the maximum point of a function f() in the given domain

Type Definitions

OdometerBE

An Iterator over a mixed-radix numbers

OdometerLE

An Iterator over a mixed-radix numbers