Trait rithmic::IndexCompress

pub trait IndexCompress {
    type T;

    fn index_compress(&self) -> IndexCompressed<'_, Self::T>;
}

Index compress a sequence of length n by mapping every value to its ordered rank

The compressed sequence is therefore guaranteed to have all its elements be integers less than n

The returned IndexCompressed implements

Deref<Target=[usize]> to access the compressed sequence

.compress(&x) to look up the rank of value x

.decompress(y) to look up the value of rank y

Examples

use rithmic::IndexCompress;

let x = [777, 9001, -40, -40, 777, -273];
let y = x.index_compress();

assert_eq!(&*y, &[2, 3, 1, 1, 2, 0]);
assert_eq!(y.compress(&-40), 1);
assert_eq!(y.decompress(2), &777)

Required Associated Types

type T

Required Methods

fn index_compress(&self) -> IndexCompressed<'_, Self::T>